Week4/django_rest/api_service/requirements.txt
asgiref==3.3.4
Django==3.2
django-filter==2.4.0
djangorestframework==3.12.4
importlib-metadata==3.10.0
Markdown==3.3.4
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
zipp==3.4.1
